\name{predict_single_tree}
\alias{predict_single_tree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Predict Outcomes Using a Single Decision Tree
}
\description{
This function takes a decision tree and a dataset as inputs and returns the predicted outcomes for each row of the dataset by traversing the tree.
}
\usage{
predict_single_tree(rf, tree, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rf}{Same randomForest object which is used below to extract tree.}
  \item{tree}{
A data frame representing the structure of the decision tree. Each row corresponds to a node in the tree and contains information about the split variable, split point, left daughter, right daughter, and terminal node status.
}
  \item{data}{
A data frame containing the input data for which predictions are to be made. Each row represents a data point, and each column corresponds to a feature.
}
}
\details{
The \code{predict_single_tree} function traverses a decision tree for each data point in the input dataset. Starting from the root node, it follows the splits defined in the tree until it reaches a terminal node, at which point it returns the prediction for that terminal node.
}
\value{
A vector of predicted outcomes, one for each row in the input data.
}
\references{
For more information on decision trees and their implementation, refer to \href{https://en.wikipedia.org/wiki/Decision_tree_learning}{Wikipedia: Decision tree learning}.
}
\author{
Markus Gruber \email{markus.gruber@tuval.at}
}
\note{
This function is intended for use with decision trees generated by specific algorithms and may need adjustments to work with different tree structures.
}